refactor(cli): extract handoff intelligence owner
